requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.9
decorator==5.1.1
Deprecated==1.2.13
Django==4.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-mssql-backend==2.8.1
django-redis==5.2.0
django-redis-admin==0.2.0
djangorestframework==3.13.1
gunicorn==20.0.4
idna==3.3
imagekitio==2.2.8
importlib-metadata==4.12.0
mssql-django==1.1.3
packaging==21.3
Pillow==8.4.0
pyodbc==4.0.32
pyparsing==3.0.9
pytz==2021.3
redis==4.3.4
requests==2.26.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.7
wrapt==1.14.1
zipp==3.8.1
